+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Show changelog after upgrade
 - Language option in settings
+- Search provider option in settings
 ### Changed
 ### Removed
 ### Fixed